+++ packages/swatch/src/swatch.ts
@@ -1,7 +1,20 @@
+export {
+	DistributionBar,
+	type DistributionBarProps,
+	type DistributionBarSegment,
+	type DistributionBarSegmentUpdate,
+	getDistributionBoundaryPercent,
+	getDistributionTotal,
+	moveDistributionBoundary,
+	removeDistributionSegment,
+	updateDistributionSegment,
+} from "./DistributionBar";
 export { getSwatchColorsBackground } from "./Swatch/SwatchColors";
 export { Swatch } from "./Swatch/SwatchRoot";
 export {
+	getSwatchSizeVariableStyle,
 	SWATCH_SHAPES,
+	SWATCH_SIZE_VALUES,
 	SWATCH_SIZES,
 	type SwatchColorStop,
 	type SwatchProps,